Add: snake movement. Changed snake position to integer

+++ include/nibbler/graphics_api.hpp
@@ -14,8 +14,8 @@ enum Key {
 };
 
 struct Position {
-    float x;
-    float y;
+    int x;
+    int y;
 };
 
 class IWindow {
